Bug 511824 - All tags are flagged as being lightweight
When listing all the tags from the repository, no checks are being
made to determine whether a given tag is annotated or lightweight.
We should identify whether a tag is annotated or not and provide this
information back to the client in the returned JSON instead of simply
flagging every tag as being lightweight.

Signed-off-by: Remy Suen <remy.suen@gmail.com>