Bug 511513 - Export non-existing folder leaks server path
The ability to export a folder obviously depends on the folder
actually existing on the server. If the file does not exist, the
native error is to of course report the full path to the client.
However, this doesn't make sense in a server setting as this would
leak the internal folder structure of the Orion workspace to the
client. We should check if the export failed because the folder
doesn't exist and to then hide the workspace's path from the user in
the returned error message if that is the case.

Signed-off-by: Remy Suen <remy.suen@gmail.com>