Bug 512061 - File locations returned from the status API are not URL encoded
When retrieving the status of a given repository, many URLs are passed
back in the JSON response to be used by the client for committing
files, looking at diffs, staging/unstaging files, and so on. These
URLs need to be URL encoded accordingly or the client will not be
able to make use of them.

Signed-off-by: Remy Suen <remy.suen@gmail.com>