Bug 512798 - URI error when deleting tag with special encoding
When a JSON response is sent back to the client regarding a tag, the
URI is encoded two times. However, when that same URI is sent back to
the server from the client, the URI is only encoded once because the
server has already performed "one" decoding itself. Orion only needs
to decode the URI once after that (instead of two times) to get at
the client's data.

Signed-off-by: Remy Suen <remy.suen@gmail.com>