Bug 512877 - URI error when deleting branch with special encoding
When the server sends a JSON response about a branch back to the
client, the URI is encoded two times. However, when that same URI is
reused to send a different request to the server, the URI is only
encoded once because the server itself has already done one decoding
pass over the URI. As a result, Orion only needs to decode the URI
once after that (instead of twice) when processing the client's
request.

Signed-off-by: Remy Suen <remy.suen@gmail.com>