Bug 513503 - Can't checkout local branch that has the same name as a tag
If a user tries to create a branch with the same name as a tag and
perform a checkout off the branch immediately, Orion will fail due to
Git's resolution prioritization of tags over branches. To fix this,
the code will now automatically prepend the `refs/heads` string to
the provided branch name to ensure that Git will find and resolve the
reference to he branch instead of the tag when trying to perform a
checkout.

Signed-off-by: Remy Suen <remy.suen@gmail.com>