Bug 510074 - Git log will not find merge commits on Node
NodeGit's RevWalk's fileHistoryWalk(*) function does not find merge
commits so the original fix for bug 507334 has to be reverted.
However, the original implementation of retrieving diffs for every
single commit to check if a file has been modified is still too
resource intensive.

Instead, the new approach of determining whether a file has been
modified is to look at the Git tree entry of the given file and
comparing it with the parent's tree entry of that file.

1. If the entry exists and the parent entry exists, compare the two
entries' ids to determine if the file has been modified or not.
2. If the entry exists and the parent entry does not exist then it
was an ADD.
3. If the entry does not exist but the parent entry exists then it
was a DELETE.

Signed-off-by: Remy Suen <remy.suen@gmail.com>