Bug 510906 - Preferences test failures
The preferences implementation uses asynchronous file I/O to write
the preferences to disk. In a testing environment, the setup and
assertions are practically run in succession. This makes it difficult
to test as the setup portion of the code for filling in the
preferences may not yet been processed when the assertions code is run.

To workaround this problem, the assertions have been placed in a
callback to be executed after a delay. This should hopefully give the
preferences implementation some time to process the setup code and
populate the preferences file before the assertions get run.

Signed-off-by: Remy Suen <remy.suen@gmail.com>