Bug 510062 - Git log graph is broken on Node
For a given file's history, the parent commit should be pointing at
the previous commit that modified the file and not the parent commit
in the repository's history.

To achieve this, for a given commit that is being filtered, we
examine the list of returned commits and see if any of its parents
are pointing at that filtered commit. If it is, then that commit will
point to the filtered commit's parent instead. This allows the chain
to be preserved as the returned commit will keep pointing at a
filtered commit's parent until it is a commit that hasn't been
filtered.

Signed-off-by: Remy Suen <remy.suen@gmail.com>