[nobug] Replace calls to bash with Node APIs
Despite Bash's popularity, not all Windows users have it installed on
their computer. Passing out file I/O calls to a shell is also a
little extraneous. Bash calls in the test code have now been replaced
by calls to Node APIs. This makes the tests run cross-platform in
addition to greatly increasing the speed in which the tests are being
completed.

Signed-off-by: Remy Suen <remy.suen@gmail.com>