Bug 508585 - Squashing unimplemented in Node
To implement `git merge --squash`, we must first retrieve the HEAD
commit and the other branch's commit.

If the other branch's commit is an ancestor of HEAD, then there is
nothing that needs to be done as we are already up-to-date.

If HEAD is an ancestor of the other branch's commit, then it's a
straightforward fast-forward operation. The resulting index should
have no conflicts. However, if the working directory and/or index has
files that are affected, the operation will abort.

If neither is an ancestor of the other, then both branches have
diverged. The resulting index may have merge conflicts. Also, like
the above case, if the working directory and/or the index has files
that overlap with the merge, the operation will abort.

Signed-off-by: Remy Suen <remy.suen@gmail.com>