Bug 510062 - Git log graph is broken on Node
When a merge occurs, sometimes changes from one side of the branch
may be completely ignored if a refactoring or rewrite has occurred.
In such a case, the merge commit and the (effectively) discarded
commits should not show up in the file's history.

To achieve this, when a merge commit is encountered during an
iteration, we'll compare the file entry's oid with the entry in the
parent commits to see if they match. If the oid does match, that
means the other branches changes have been discarded. If such a
discard has taken place, the merge commit itself and all the commits
from those discarded branches will be flagged to be ignored so that
they are not included in the returned JSON response.

Signed-off-by: Remy Suen <remy.suen@gmail.com>