[node] Tagging a commit with an existing tag name causes two responses to be sent back
In the tagging code, the promise chain uses a done(*) function for
sending back the response. This causes an issue when an error occurs
such as when the user tries to tag a commit with a name that's
already in use. The catch(*) function will resolve and send an error
response back and then the done(*) function will also resolve and try
to send a response back. The code should instead be structured such
that a success response should only be sent back if everything is
okay. It should not be sending a success response back irregardless
of whether the commit was tagged or not.

Signed-off-by: Remy Suen <remy.suen@gmail.com>