Bug 515613 - [node] Can't merge two different histories together
When a merge is attempted, a common ancestor is required as it may
simply be a fast-forward or the active branch may already be
up-to-date depending on what the tip commit of the other branch is.
However, if the user tries to merge an unrelated history, there will
not be a common ancestor between the two branches. We need to handle
this case and immediately try to merge the two branches instead.

Signed-off-by: Remy Suen <remy.suen@gmail.com>